requirements.txt
aiohttp
websockets
numpy
pycairo
PyGObject==3.46.0
